Move network-msg-processing code out of main to its own file
>>> based on bitcoin/bitcoin@e736772c56a883e2649cc8534dd7857a0718ec56
